Project 1 updates

+++ compare.sh
@@ -1,20 +1,46 @@
 #!/usr/bin/bash
 
-# Runs your code with the ./tests/text.in as input
-# diffs with tests/test.ans. Writes diff to diff.txt
-diff <(run.sh ./tests/test.in) <(cat ./tests/test.ans) > ./diff.txt
+failedTests=0
+for inFile in tests/*.in; do
 
-# Check if there is any diff in the files
-if [ $(cat ./diff.txt | wc --bytes) -gt 0 ]; then
-    # If there is a diff, print the result of the diff
-    echo "FAILED"
-    echo "Output differs... diff follows:"
-    cat ./diff.txt
+    base=`echo "$inFile" | cut -f 1 -d '.'`
+    testName=`echo "$base" | cut -d '/' -f 2`
+    ansFile="$base.ans"
+
+    if [[ -f $ansFile ]]; then
+        
+        echo ""
+        echo "Starting Test: '$testName'"
+        echo "----------------------------------"
+        
+        # Runs your code with the ./tests/text.in as input
+        # diffs with tests/test.ans. Writes diff to diff.txt
+        diff <(run.sh $inFile) <(cat $ansFile) > ./diff.txt
+
+        # Check if there is any diff in the files
+        if [ $(cat ./diff.txt | wc --bytes) -gt 0 ]; then
+            # If there is a diff, print the result of the diff
+            ((failedTests+=1))
+            echo "FAILED"
+            echo "Output differs... diff follows:"
+            cat ./diff.txt
+        else
+            # If they're the same, output such
+            echo "PASSED"
+            echo "`cat $ansFile | wc -l` lines of test output match!"
+        fi
+
+        # Remove the temp diff file
+        rm ./diff.txt
+    else
+        echo "Cannot find associated answer file: $ansFile"
+    fi
+done
+
+echo ""
+if [[ failedTests -eq 0 ]]; then
+    echo "Passed ALL tests!"
 else
-    # If they're the same, output such
-    echo "PASSED"
-    echo "`cat ./tests/test.ans | wc -l` lines of test output match!"
+    echo "Did not pass all tests ($failedTests failed). See above for details"
 fi
-
-# Remove the temp diff file
-rm ./diff.txt
+echo ""
